+++ docs/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 #### 7.3.2-SNAPSHOT
 
-* PR [#270](https://github.com/azagniotov/stubby4j/pull/270) - Additive proxy strategy that sets additional headers (https://github.com/azagniotov)
+* PR [#270](https://github.com/azagniotov/stubby4j/pull/270) - Added `additive` proxy strategy behavior that sets additional headers to the request (https://github.com/azagniotov)
 
 #### 7.3.1